Refactor to move syntax config to lexer
